Refactor SyncReporter related types into separate module
Cut the size of sync.rs in half.
