+++ crates/spk-storage/src/storage/archive.rs
@@ -194,7 +194,7 @@ where
     let components = src_repo.read_components(pkg).await?;
     tracing::info!(%pkg, "exporting");
     let syncer = spfs::Syncer::new(src_repo.spfs_repository_handle(), dst_repo)
-        .with_reporter(spfs::sync::SyncReporters::console());
+        .with_reporter(spfs::sync::reporter::SyncReporters::console());
     let desired = components.iter().map(|i| *i.1).collect();
     syncer.sync_env(desired).await?;
     dst_repo.publish_package(&spec, &components).await?;
